Move pointwise_scatter optimization to joint_graph stage from post_grad, fixes #129449
